+++ src/libs/auth/src/openid/jkwset/impls.rs
@@ -1,6 +1,6 @@
-use crate::openid::jwt::types::errors::{JwtFindKidError};
-use crate::openid::jwtset::types::errors::GetOrRefreshJwksError;
-use crate::openid::jwtset::types::interface::GetOpenIdCertificateArgs;
+use crate::openid::jkwset::types::errors::GetOrRefreshJwksError;
+use crate::openid::jkwset::types::interface::GetOpenIdCertificateArgs;
+use crate::openid::jwt::types::errors::JwtFindKidError;
 use crate::openid::types::provider::OpenIdProvider;
 
 impl From<JwtFindKidError> for GetOrRefreshJwksError {
