+++ src/libs/auth/src/openid/jkwset/jwks.rs
@@ -1,8 +1,8 @@
+use crate::openid::jkwset::constants::REFRESH_COOLDOWN_NS;
+use crate::openid::jkwset::fetch::fetch_openid_certificate;
+use crate::openid::jkwset::types::errors::GetOrRefreshJwksError;
 use crate::openid::jwt::types::cert::Jwks;
 use crate::openid::jwt::unsafe_find_jwt_kid;
-use crate::openid::jwtset::constants::REFRESH_COOLDOWN_NS;
-use crate::openid::jwtset::fetch::fetch_openid_certificate;
-use crate::openid::jwtset::types::errors::GetOrRefreshJwksError;
 use crate::openid::types::provider::OpenIdProvider;
 use crate::state::types::state::OpenIdCachedCertificate;
 use crate::state::{cache_certificate, get_cached_certificate, record_fetch_attempt};
