+++ src/libs/auth/src/openid/mod.rs
@@ -1,6 +1,6 @@
 mod impls;
+pub mod jkwset;
 pub mod jwt;
-pub mod jwtset;
 pub mod types;
 mod utils;
 mod verify;
